style: split out program metrics into another module
