+++ program-runtime/src/lib.rs
@@ -11,6 +11,7 @@ pub mod invoke_context;
 pub mod loaded_programs;
 pub mod mem_pool;
 pub mod memory;
+pub mod program_metrics;
 pub mod serialization;
 pub mod stable_log;
 pub mod sysvar_cache;
